engine/requirements.txt
starlette>=0.41.0
tzdata>=2024.1
uvicorn[standard]==0.35.0
motor==3.7.1
httpx==0.28.1
python-dotenv==1.1.1
tenacity==9.1.2
pytest==8.4.1
pytest-asyncio==0.24.0
